user.voice_duration is now an unsigned bigint (from ~ 24 days for a signed int to more than 500 million years for an unsigned bigint)

+++ assets/sql/init.sql
@@ -74,7 +74,7 @@ create table if not exists USER (
     store_points int default 0 not null,
     leaderboard_points int default 0 not null,
     messages_sent int default 0 not null,
-    voice_duration int default 0 not null,
+    voice_duration bigint unsigned default 0 not null,
     nitro_boost tinyint(1) default 0 not null,
     bump_count int default 0 not null,
     primary key (user_id, guild_id),
